requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
chardet==4.0.0
click==8.1.7
dataclasses==0.6
deprecation==2.1.0
fastapi==0.110.1
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
gotrue==2.4.2
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.26.0
idna==2.10
iniconfig==2.0.0
oauthlib==3.2.2
packaging==24.0
pluggy==1.4.0
postgrest==0.16.3
proto-plus==1.23.0
protobuf==4.25.3
py==1.11.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.0
pydantic_core==2.18.1
pyparsing==3.1.2
pytest==6.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==1.0.4
realtime-py==0.1.3
requests==2.25.1
requests-oauthlib==2.0.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
storage3==0.7.4
StrEnum==0.4.15
supabase==2.4.2
supafunc==0.4.5
toml==0.10.2
tqdm==4.66.2
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.29.0
websockets==12.0
